Allow to configure multiple backends
When defining a Daemonset manually, one has the (very toil-intensive)
option to configure multiple backends to which the agent should report
as documented at [1]. However, there is no way to add configmap entries,
volumes and volume mounts to the `instana-agent` container in the current
Helm chart, so agents installed via Helm are limited to report to one
backend.

This change introduces an easy way to set up additional backends to which
the Instana agent will report to.

[1] https://www.instana.com/docs/setup_and_manage/host_agent/configuration/#configure-multiple-backends